modules/imagedownloader/pyproject.toml
[tool.poetry]
name = "imagedownloader"
version = "0.1.0"
description = ""
authors = ["Benjamin Beeker <benjamin.beeker@intusurg.com>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
requests = "^2.26.0"
google-images-download = {path = "../google-images-download", develop = true }
pandas = "^1.3.4"
Unidecode = "^1.3.2"
aiohttp = "^3.7.4"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
